Major level / save-format changes
I changed the new save-format to load/save tile ids instead of a table with the tiles variables.
I also converted and migrated the existing png levels to the new format and made quite some changes to the editor to fix some stuff that bugged me from the beginning.
Also it's now possible to edit the "campaign" levels and play them, because levels now get saved individually in the appdata save directory and automatically replace the default levels if there's one with the same id.
Note: there still seems to be a bug that reduces the framerate significantly (when vsync is off). I'll look into that.