Fixed levels / bug fixes
Thanks to the new editor and save-format I was able to easily edit all existing levels to match the new tileset. I also managed to fix a bug where the bomb teleported randomly when the player spawned it into a wall.
The performance regression from the save-format update also has been resolved.